requirements.txt
numpy>=1.18.5
matplotlib>=3.3.2
pillow>=7.2.0